# docs/swagger.yaml
basePath: /
definitions:
  dtos.APIError:
    properties:
      error:
        type: string
    type: object
  dtos.Job:
    properties:
      descricao:
        type: string
      localizacao:
        type: string
      'posted:':
        type: string
      titulo:
        type: string
    type: object
  dtos.Message:
    properties:
      message:
        type: string
    type: object
  dtos.User:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8001
info:
  contact: {}
  description: API do Sistema de Recrutamento e Seleção Inteligente (SIRSI)
  title: API de Autenticação e Tarefas SIRSI
  version: "1.0"
paths:
  /cadastrar/vagas:
    post:
      consumes:
      - application/json
      description: Registra uma novo vaga no banco
      parameters:
      - description: Dados da vaga
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/dtos.Job'
      produces:
      - application/json
      responses:
        "201":
          description: Vaga criado com sucesso
          schema:
            $ref: '#/definitions/dtos.Message'
        "400":
          description: Erro de validação
          schema:
            $ref: '#/definitions/dtos.APIError'
      summary: Criar uma nova vaga
      tags:
      - Create Jobs
  /jobs:
    get:
      consumes:
      - application/json
      description: Retorna todas as vagas diponíveis no Banco (Colletion Testes)
      produces:
      - application/json
      responses:
        "201":
          description: Vagas retornadas com sucesso!
          schema:
            $ref: '#/definitions/dtos.Message'
        "400":
          description: Erro ao retornar as vagas
          schema:
            $ref: '#/definitions/dtos.APIError'
      summary: Retorna vagas
      tags:
      - jobs
  /users:
    post:
      consumes:
      - application/json
      description: Registra um novo usuário na API
      parameters:
      - description: Dados do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dtos.User'
      produces:
      - application/json
      responses:
        "201":
          description: Usuário criado
          schema:
            $ref: '#/definitions/dtos.Message'
        "400":
          description: Erro de validação
          schema:
            $ref: '#/definitions/dtos.APIError'
      summary: Criar um novo usuário
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Value: Bearer abc... (Bearer+space+token)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"